requirements.txt
psutil==5.0.1
python-osc==1.6.2
pySerial==3.2.1
